+++ src/lib/container.js
@@ -4,22 +4,26 @@ export const createContainer = option => {
   return { ...option, result: postcss.root() };
 };
 
-export const createUpdaterFn = containers => atRule => {
-  const killRules = [];
-
-  containers.forEach(container => {
-    container.match.forEach(regex => {
-      const previouslyFound = findRegex(regex, killRules);
-      if (!previouslyFound && regex.test(atRule.params)) {
-        container.result.append(atRule.clone());
-        killRules.push(atRule);
-      }
-    });
-  });
-
-  killRules.forEach(rule => rule.remove());
+export const matchAll = (arr, str) => {
+  return arr.filter(regex => {
+    return regex.test(str)
+  }).length === arr.length;
 };
 
-export const findRegex = (regex, arr) => {
-  return arr.find(elem => regex.toString() === elem.toString());
+export const matchAny = (arr, str) => {
+  return !!arr.find(regex => {
+    return regex.test(str)
+  });
+};
+
+export const createUpdaterFn = containers => atRule => {
+  const killRules = [];
+  containers.forEach(({ skip, match, result }) => {
+    if (matchAny(skip, atRule)) return;
+    if (matchAny(match, atRule)) {
+      result.append(atRule.clone());
+      killRules.push(atRule);
+    }
+  });
+  killRules.forEach(rule => rule.remove());
 };