+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.3.0
+- Added:
+  - option: `files.skip` to override matched media queries
+- Updated:
+  - option: `outpath` will resolve to `process.cwd()` if not otherwise specified
+
 ## 1.1.0 - 2017.10.16
 - Captures media queries in multiple files
 - Improved testing